Updated install scripts for linux.

+++ uninstall.sh
@@ -1,26 +1,25 @@
 #!/bin/bash
 
-dotfilesDir="~/.dotfiles"
-files=".fonts .zshrc .gitconfig .tmux.conf .clang-format"
-oldFiles=${dotfilesDir}/oldDotfiles
+DOTFILES_DIR="${HOME}/.dotfiles"
+FILES=".SpaceVim.d .fonts .zshrc .gitconfig .tmux.conf .clang-format"
+OLD_DOTFILES_DIR=${DOTFILES_DIR}/OldDotFiles
 
-cd ${HOME}
+if [[ -d ${OLD_DOTFILES_DIR} ]]; then
 
-echo "Removing all dotfiles within your home directory..."
-rm -rf ~/.oh-my-zsh
-for file in ${files}; do
-	echo "Deleting ${file} ..."
-	rm -rf ${file}
-	echo "done"
-done
-echo "done"
-				
-echo "Reverting back to your previous settings..."
-cd ${oldFiles}
-for file in ${oldFiles}; do
-	mv ${file} ${HOME}/${file}
-done
-
-cd ${HOME}
-rm -rf ${dotfilesDir}
-echo "Uninstall complete!"
+    echo "Removing all dotfiles within your home directory..."
+    rm -rf ~/.oh-my-zsh
+    for file in ${FILES}; do
+    	rm -rf ${HOME}/${file}
+    done
+    				
+    echo "Reverting back to your previous settings..."
+    for file in ${OLD_DOTFILES_DIR}; do
+    	mv ${OLD_DOTFILES_DIR}/${file} ${HOME}/${file}
+    done
+    
+    #todo: uninstall spacevim?
+    
+    cd ${HOME}
+    rm -rf ${DOTFILES_DIR}
+    echo "Uninstall complete!"
+fi
